Minor Commit
Added a small box to the right which will display work description of each timeframe when respective button is pressed.
Added buttons which will change state of Description box text when i implement the code.
Made a few things easier to understand and split a couple of component pieces into new files, next step is getting the description box on the right to render correctly and display data instead of being empty...